Moved MOS6507Status type definition into a separate file.
